+++ tools/integration_tests/operations/create_three_level_dir_test.go
@@ -145,5 +145,5 @@ func TestCreateThreeLevelDirectories(t *testing.T) {
 	}
 
 	// Delete objects from mountedDirectory after testing.
-	operations.CleanMntDir()
+	clean_mount_dir.CleanMntDir()
 }
