+++ tools/integration_tests/util/setup/implicit_and_explicit_dir_setup/implicit_and_explicit_dir_setup.go
@@ -91,5 +91,5 @@ func CreateExplicitDirectory(t *testing.T) {
 	}
 
 	// Closing file at the end.
-	operations.CloseFile(file)
+	defer operations.CloseFile(file)
 }
